refactor: 클래스명 변경 및 검증 메소드 로직 변경
