+++ backend/src/main/java/hanglog/image/dto/ImagesResponse.java
@@ -1,6 +1,6 @@
 package hanglog.image.dto;
 
-import static hanglog.image.util.ImageNameUrlConverter.convertNameToUrl;
+import static hanglog.image.util.ImageUrlConverter.convertNameToUrl;
 
 import hanglog.image.domain.Image;
 import java.util.List;
@@ -18,7 +18,7 @@ public class ImagesResponse {
     public static ImagesResponse of(final List<Image> images) {
         return new ImagesResponse(
                 images.stream()
-                        .map(image -> convertNameToUrl(image.getImageName()))
+                        .map(image -> convertNameToUrl(image.getName()))
                         .toList());
     }
 }
